requirements.txt
aiobotocore==2.5.4
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.12.0
anyio==3.7.1
apprise==1.5.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
botocore==1.31.17
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.4.1
cryptography==41.0.4
dask==2023.9.2
dateparser==1.1.8
distributed==2023.9.2
docker==6.1.3
duckdb==0.9.0
fastapi==0.103.1
frozenlist==1.4.0
fsspec==2023.9.2
google-auth==2.23.0
google-auth-oauthlib==1.1.0
graphviz==0.20.1
greenlet==2.0.2
griffe==0.36.2
gspread==5.11.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kubernetes==28.1.0
locket==1.0.0
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
markdown2==2.4.10
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.6
multidict==6.0.4
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.9.7
packaging==23.1
pandas==2.1.1
partd==1.4.1
pathspec==0.11.2
pendulum==2.1.2
prefect==2.13.2
prefect-dask==0.2.5
psutil==5.9.5
psycopg==3.1.10
psycopg-binary==3.1.10
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-frontmatter==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3fs==2023.9.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
starlette==0.27.0
tblib==2.0.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.23.2
websocket-client==1.6.3
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2
zict==3.0.0
zipp==3.17.0
